feat: issue jwt with TempToken
